Rename handlers, merge transform, rename shared type
Rename index.ts → {handler-name}.ts and merge transform.ts
into each handler file for both staff-assignment and trustee-
appointment handlers. Rename CmmapStagingRow → CmmapCamsRow
and move to cmmap-cams-row.ts. Rename seed file from
02-seed-cmmap-staging.sql → 02-seed-cmmap-cams.sql. Update
READMEs to remove stale references to CMMAP_STAGING and the
separate cams-downstream database concept.

Jira ticket: CAMS-616
