+++ downstream/index.ts
@@ -1,2 +1,2 @@
-import './staff-assignment-handler';
-import './trustee-appointment-handler';
+import './staff-assignment-handler/staff-assignment-handler';
+import './trustee-appointment-handler/trustee-appointment-handler';
