+++ downstream/trustee-appointment-handler/trustee-appointment-handler.ts
@@ -1,11 +1,11 @@
 import { app, InvocationContext, output } from '@azure/functions';
 import * as sql from 'mssql';
 import { buildQueueName } from '@common/queues';
-import {
-  TrusteeAppointmentDownstreamEvent,
-  transformTrusteeToStagingRow,
-  CmmapStagingRow,
-} from './transform';
+import { TrusteeAppointmentDownstreamEvent } from '@common/cams/dataflow-events';
+import { CmmapCamsRow } from '../shared/cmmap-cams-row';
+
+export type { TrusteeAppointmentDownstreamEvent };
+export type { CmmapCamsRow };
 
 const MODULE_NAME = 'TRUSTEE-APPOINTMENT-HANDLER';
 const QUEUE_NAME = buildQueueName('TRUSTEE-APPOINTMENT-EVENT');
@@ -28,7 +28,82 @@ function getSqlConfig(): sql.config {
   };
 }
 
-async function upsertTrusteeAppointment(row: CmmapStagingRow): Promise<void> {
+export function parseCaseId(caseId: string): { div: number; year: number; number: number } {
+  const match = caseId.match(/^(\d{3})-(\d{2})-(\d{5})$/);
+  if (!match) {
+    throw new Error(`Invalid CAMS case ID format: ${caseId}`);
+  }
+  return {
+    div: parseInt(match[1], 10),
+    year: parseInt(match[2], 10),
+    number: parseInt(match[3], 10),
+  };
+}
+
+// Convert ISO date string to ACMS positional integer date format (YYYYMMDD)
+export function toAcmsDateNumeric(isoDateString: string): number {
+  const datePortion = isoDateString.split('T')[0];
+  return parseInt(datePortion.replace(/-/g, ''), 10);
+}
+
+// Parse "{GROUP_DESIGNATOR}-{PROF_CODE}" (e.g. "NY-00063") into components
+function parseProfessionalId(acmsProfessionalId: string): { group: string; code: number } {
+  const dashIndex = acmsProfessionalId.indexOf('-');
+  const group = acmsProfessionalId.slice(0, dashIndex);
+  const code = parseInt(acmsProfessionalId.slice(dashIndex + 1), 10);
+  return { group, code };
+}
+
+export function transformTrusteeToStagingRow(
+  event: TrusteeAppointmentDownstreamEvent,
+): CmmapCamsRow {
+  const { div, year, number } = parseCaseId(event.caseId);
+  const { group, code } = parseProfessionalId(event.acmsProfessionalId);
+  const now = new Date();
+
+  const isUnassigned = !!event.unassignedOn;
+  const apptDateSource = event.appointedDate ?? event.assignedOn;
+  const apptDate = toAcmsDateNumeric(apptDateSource);
+  const dispDate = isUnassigned ? toAcmsDateNumeric(event.unassignedOn!) : null;
+
+  return {
+    DELETE_CODE: ' ',
+    CASE_DIV: div,
+    CASE_YEAR: year,
+    CASE_NUMBER: number,
+    RECORD_SEQ_NBR: 1,
+    PROF_CODE: code,
+    GROUP_DESIGNATOR: group,
+    APPT_TYPE: 'TR',
+    APPT_DATE: apptDate,
+    APPT_DATE_DT: new Date(apptDateSource),
+    APPT_DISP: isUnassigned ? 'WD' : 'GR',
+    DISP_DATE: dispDate,
+    DISP_DATE_DT: isUnassigned ? new Date(event.unassignedOn!) : null,
+    COMMENTS: null,
+    APPTEE_ACTIVE: isUnassigned ? 'N' : 'Y',
+    ALPHA_SEARCH: null,
+    USER_ID: 'CAMS',
+    HEARING_SEQUENCE: null,
+    REGION_CODE: null,
+    RGN_CREATE_DATE: null,
+    RGN_UPDATE_DATE: null,
+    RGN_CREATE_DATE_DT: null,
+    RGN_UPDATE_DATE_DT: null,
+    CDB_CREATE_DATE: toAcmsDateNumeric(now.toISOString()),
+    CDB_UPDATE_DATE: toAcmsDateNumeric(now.toISOString()),
+    CDB_CREATE_DATE_DT: now,
+    CDB_UPDATE_DATE_DT: now,
+    UPDATE_DATE: now,
+    SOURCE: 'CAMS',
+    CAMS_CASE_ID: event.caseId,
+    CAMS_USER_ID: 'CAMS',
+    CAMS_USER_NAME: 'CAMS',
+    LAST_UPDATED: now,
+  };
+}
+
+async function upsertTrusteeAppointment(row: CmmapCamsRow): Promise<void> {
   const pool = await sql.connect(getSqlConfig());
   const request = pool.request();
